third_party/WebKit/LayoutTests/inspector/inspector-backend-commands-expected.txt

Tests correctness of promisified protocol commands.

error: Request with id = ## failed. {"message":"this is the error message"}
commandError: then result: ["this is the error message"]
commandArgs0: then result: [null]
error: Protocol Error: Optional callback argument for method 'Profiler.commandArgs0' call must be a function but its type is 'number'.
commandArgs0: catch result: {}
commandArgs1Rets0: then result: [null]
error: Protocol Error: Optional callback argument for method 'Profiler.commandArgs1Rets0' call must be a function but its type is 'number'.
commandArgs1Rets0: catch result: {}
error: Protocol Error: Invalid type of argument 'arg1' for method 'Profiler.commandArgs1Rets0' call. It must be 'number' but it is 'string'.
commandArgs1Rets0: catch result: {}
error: Protocol Error: Invalid number of arguments for method 'Profiler.commandArgs1Rets0' call. It must have the following arguments '[{"name":"arg1","type":"number","optional":false}]'.
commandArgs1Rets0: catch result: {}
commandArgs1Rets1: then result: [null,{"value1":234}]
commandArgs1Rets1: then result: [null,{"value1":234}]
commandArgs3Rets3: then result: [null,{},345,"alph"]
commandArgs3Rets3: then result: [null,{},345,null]
commandArgs3Rets3: then result: [null,{},null,"alph"]
error: Protocol Error: Invalid type of argument 'arg2' for method 'Profiler.commandArgs3Rets3' call. It must be 'number' but it is 'string'.
commandArgs3Rets3: catch result: {}
commandArgs3Rets3: then result: [null,{},null,null]
error: Protocol Error: Invalid number of arguments for method 'Profiler.commandArgs3Rets3' call. It must have the following arguments '[{"name":"arg1","type":"object","optional":false},{"name":"arg2","type":"number","optional":true},{"name":"arg3","type":"string","optional":true}]'.
commandArgs3Rets3: catch result: {}
error: Protocol Error: Invalid type of argument 'arg2' for method 'Profiler.commandArgs3Rets3' call. It must be 'number' but it is 'string'.
commandArgs3Rets3: catch result: {}
error: Protocol Error: Invalid type of argument 'arg2' for method 'Profiler.commandArgs3Rets3' call. It must be 'number' but it is 'string'.
commandArgs3Rets3: catch result: {}
